refactor(extension): co-locate redirect rules + model entries into per-site adapters
Move the redirect-rule database out of @movar/host-match (now predicates-only)
into apps/extension/src/sites/<site>/, one folder per site:
- index.ts: the eager redirect rule + its value maps (YT_GL, DDG_REGION) and,
  for sites with an extractor, the lazy model-chunk descriptor;
- model.ts: the lazy page-content model chunk entry (moved from dynamic/models/).
A tiny eager registry (sites/registry.ts) answers getRuleForHost + resolveModelChunk.

@movar/host-match keeps only isGoogleHost/isYouTubeHost — the predicates shared
with the page-content extractors. SiteRule/LangStrategy/encodedValue move to
sites/types.ts (extension-internal). GOOGLE_LR collapses into a prefix: 'lang_'
on the lr param (behaviour-identical; the value map only carries real
per-language data now). Build wiring (wxt CAPABILITY_ENTRY_POINTS, .fallowrc)
and the import-graph boundary guard follow the moved model entries.

Behaviour-preserving: host-match 18 + extension 884 tests green; chunk budgets
unchanged.
