+++ apps/extension/src/lib/correction-mechanism.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import type { SiteRule } from '@movar/host-match';
+import type { SiteRule } from '../sites/types';
 import { mechanismForStrategy } from './correction-mechanism';
 
 const rule = (strategy: SiteRule['strategy']): SiteRule => ({ match: 'example.com', strategy });
